fix(ui): validate newly entered redaction sentinels
