+++ ui/lib/utils/validation.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { isRedacted } from "./validation";
+import { getPasswordPolicyFailures, isRedacted } from "./validation";
 
 describe("isRedacted", () => {
 	it.each(["<redacted>", "<REDACTED>", "[redacted]", "[REDACTED]"])("recognizes the backend sentinel %s", (value) => {
@@ -8,5 +8,22 @@ describe("isRedacted", () => {
 
 	it("does not classify a real password as redacted", () => {
 		expect(isRedacted("StrongPassword1!")).toBe(false);
+	});
+});
+
+describe("getPasswordPolicyFailures", () => {
+	it.each([
+		["<redacted>", ["at least 12 characters", "one uppercase letter", "one number"]],
+		["[REDACTED]", ["at least 12 characters", "one lowercase letter", "one number"]],
+	])("validates a newly entered sentinel %s", (password, expectedFailures) => {
+		expect(getPasswordPolicyFailures(password, false)).toEqual(expectedFailures);
+	});
+
+	it.each(["<redacted>", "[REDACTED]"])("skips an unchanged server sentinel %s", (password) => {
+		expect(getPasswordPolicyFailures(password, true)).toEqual([]);
+	});
+
+	it("accepts a strong newly entered password", () => {
+		expect(getPasswordPolicyFailures("StrongPassword1!", false)).toEqual([]);
 	});
 });
